Refactor argument handling and fix import paths in pipeline scripts for improved functionality
