+++ main-pgse.py
@@ -1,7 +1,10 @@
-from pgse.enviromnet import args
+from pgse.environment.args import get_parser
 from pgse import TrainingPipeline
 
 if __name__ == "__main__":
+    parser = get_parser()
+    args = parser.parse_args()
+
     pipeline = TrainingPipeline(
         args.data_dir,
         args.label_file,
@@ -21,5 +24,4 @@ if __name__ == "__main__":
         args.nodes,
         args.workers
     )
-
     pipeline.run()
