+++ pgse/__init__.py
@@ -2,4 +2,6 @@ from pgse.pipeline.pgse_pipeline import Pipeline as TrainingPipeline
 from pgse.pipeline.pgse_inference_pipeline import Pipeline as InferencePipeline
 from pgse.pipeline.regular_pipline import Pipeline as PureXGBPipeline
 
-__all__ = ['TrainingPipeline', 'InferencePipeline', 'PureXGBPipeline']
+TrainingPipeline = TrainingPipeline
+InferencePipeline = InferencePipeline
+PureXGBPipeline = PureXGBPipeline
