--- pgse/enviromnet/__init__.py
@@ -1,1 +1,0 @@
-from pgse.enviromnet.args import args
